Move nav-only components to components/navigation
(cherry picked from commit 73cba13)
